requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
supabase==1.0.3
python-dotenv==1.0.0
openai==1.3.0
PyJWT==2.8.0
Werkzeug==2.3.7 